[ALPS05255985] cmdq: fix build error for 32-bit project
1. Fix build error for 32-bit project
2. Check return code of copy_from_user

MTK-Commit-Id: c26d5696cfee0c44a400cc2684fdffada958cea6

CR-Id: ALPS05255985
Feature: [Module]CMDQ Driver

+++ drivers/misc/mediatek/cmdq/v3/cmdq_driver.c
@@ -419,7 +419,8 @@ static long cmdq_driver_create_secure_medadata(
 	/* always clear to prevent free unknown memory */
 	pCommand->secData.addrMetadatas = 0;
 	for (i = 0; i < ARRAY_SIZE(pCommand->secData.ispMeta.ispBufs); i++) {
-		isp_bufs[i] = (void *)pCommand->secData.ispMeta.ispBufs[i].va;
+		isp_bufs[i] = (void *)(unsigned long)
+			pCommand->secData.ispMeta.ispBufs[i].va;
 		pCommand->secData.ispMeta.ispBufs[i].va = 0;
 	}
